Fix a 32-bit compilation warning.
Not sure if we can use %zu on all platforms we support (it's Posix),
but if so that would be a much better way to do it.

+++ crawl-ref/source/evoke.cc
@@ -636,7 +636,7 @@ string manual_skill_names(bool short_text)
     if (short_text && skills.size() > 1)
     {
         char buf[40];
-        sprintf(buf, "%lu skills", skills.size());
+        sprintf(buf, "%lu skills", (unsigned long) skills.size());
         return string(buf);
     }
     else
